1.15.7: hitbox fix for extractors
