+++ src/main/java/com/lothrazar/cyclicmagic/block/pump/energy/BlockEnergyPump.java
@@ -24,28 +24,21 @@
 package com.lothrazar.cyclicmagic.block.pump.energy;
 
 import com.lothrazar.cyclicmagic.IHasRecipe;
-import com.lothrazar.cyclicmagic.core.block.BlockBaseFacingOmni;
+import com.lothrazar.cyclicmagic.block.pump.BlockPump;
 import com.lothrazar.cyclicmagic.core.registry.RecipeRegistry;
 import com.lothrazar.cyclicmagic.gui.ForgeGuiHandler;
 import net.minecraft.block.ITileEntityProvider;
-import net.minecraft.block.material.Material;
 import net.minecraft.init.Blocks;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.crafting.IRecipe;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.world.World;
 
-public class BlockEnergyPump extends BlockBaseFacingOmni implements ITileEntityProvider, IHasRecipe {
+public class BlockEnergyPump extends BlockPump implements ITileEntityProvider, IHasRecipe {
 
   public BlockEnergyPump() {
-    super(Material.WOOD);
-    this.setHardness(3F);
-    this.setResistance(3F);
-    this.setHarvestLevel("pickaxe", 1);
-    this.setTranslucent();
-    //  setLightOpacity(0);
+    super();
     super.setGuiId(ForgeGuiHandler.GUI_INDEX_ENERGYPUMP);
-    this.placeType = PlacementType.SIDE_BLOCK;
   }
 
   @Override
